+++ apps/docs/assets/bundled-scripts/cache-assets.js
@@ -1,0 +1,13 @@
+(async function() {
+	const active = 'ACTIVE';
+	if (active === 'false') {
+		return;
+	}
+	const registration = await navigator.serviceWorker.register(
+		'/sw.js',
+		{
+			scope: '/',
+		}
+	);
+})();
+