+++ libs/components/src/lib/nav-item/nav-item.ts
@@ -1,4 +1,5 @@
 import { applyMixins } from '@microsoft/fast-foundation';
+import { html } from '@microsoft/fast-element';
 import { AffixIcon } from '../../shared/patterns/affix';
 import { TextAnchor } from '../text-anchor/text-anchor';
 
@@ -8,8 +9,13 @@ import { TextAnchor } from '../text-anchor/text-anchor';
  * Based largely on the {@link https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a | <a> element }.
  *
  * @public
+ * @slot meta - A slot to add additional content to the nav item.
  */
-export class NavItem extends TextAnchor {}
+export class NavItem extends TextAnchor {
+	override getBodyTemplate() {
+		return html<NavItem>`<slot name="meta"></slot>`;
+	}
+}
 
 export interface NavItem extends AffixIcon {}
 applyMixins(NavItem, AffixIcon);