+++ libs/components/src/lib/option/option.spec.ts
@@ -32,7 +32,11 @@ describe('vwc-option', () => {
 	});
 
 	describe('icon', () => {
-		it('should add an icon to the option', async () => {
+		it('should have an icon slot', async () => {
+			expect(Boolean(element.shadowRoot?.querySelector('slot[name="icon"]'))).toEqual(true);
+		});
+
+		it('should have an icon when icon is set without slotted icon', async () => {
 			element.icon = 'home';
 			await elementUpdated(element);