+++ libs/components/src/lib/tab/tab.ts
@@ -2,12 +2,47 @@ import { attr } from '@microsoft/fast-element';
 import { applyMixins, Tab as FoundationTab } from '@microsoft/fast-foundation';
 import { AffixIconWithTrailing } from '../../shared/patterns/affix';
 
+import type { Connotation, Shape} from '../enums.js';
+
+/**
+ * Types of tab connotation.
+ *
+ * @public
+ */
+export type TabConnotation = Extract<Connotation,
+| Connotation.Accent
+| Connotation.CTA>;
+
+/**
+ * Types of avatar shape.
+ *
+ * @public
+ */
+export type TabShape = Extract<Shape, Shape.Rounded | Shape.Sharp>;
+
 /**
  * Base class for tab
  *
  * @public
  */
 export class Tab extends FoundationTab {
+	/**
+	 * The connotation the tab should have.
+	 *
+	 * @public
+	 * @remarks
+	 * HTML Attribute: connotation
+	 */
+	@attr connotation?: TabConnotation;
+	/**
+	 * The shape the tab should have.
+	 *
+	 * @public
+	 * @remarks
+	 * HTML Attribute: shape
+	 */
+	@attr shape?: TabShape;
+
 	/**
 	 * Indicates the tab's label.
 	 *