+++ libs/components/src/lib/tabs/tabs.template.ts
@@ -1,12 +1,12 @@
-
 import { html, ref, slotted, when } from '@microsoft/fast-element';
 import { classNames } from '@microsoft/fast-web-utilities';
 import type { Tabs } from './tabs.js';
 
 const getClasses = ({
-	orientation
+	connotation, orientation
 }: Tabs) => classNames(
 	'base',
+	[`connotation-${connotation}`, Boolean(connotation)],
 	[`orientation-${orientation}`, Boolean(orientation)],
 );
 
@@ -18,18 +18,20 @@ const getClasses = ({
  */
 export function TabsTemplate<T extends Tabs>() {
 	return html<T>`
-	<template role="tablist">
-		<div class="${getClasses}">
-			<div class="tablist" ${ref('tablist')}>
-				<slot name="tab" ${slotted('tabs')}></slot>
-				${when(x => x.showActiveIndicator, html<T>`
-					<div ${ref('activeIndicatorRef')} class="active-indicator"></div>
-				`)}
+		<template role="tablist">
+			<div class="${getClasses}">
+				<div class="tablist-wrapper">
+					<div class="tablist" ${ref('tablist')}>
+						<slot name="tab" ${slotted('tabs')}></slot>
+						${when(x => x.showActiveIndicator, html<T>`
+							<div ${ref('activeIndicatorRef')} class="active-indicator"></div>
+						`)}
+					</div>
+				</div>
+				<div class="tabpanel">
+					<slot name="tabpanel" ${slotted('tabpanels')}></slot>
+				</div>
 			</div>
-			<div class="tabpanel">
-				<slot name="tabpanel" ${slotted('tabpanels')}></slot>
-			</div>
-		</div>
-	</template>
-    `;
+		</template>
+	`;
 }