+++ libs/components/src/lib/tag-group/tag-group.ts
@@ -4,6 +4,7 @@ import { FoundationElement } from '@microsoft/fast-foundation';
  * Base class for tag-group
  *
  * @public
+ * @slot - The content of the tag-group.
  */
 export class TagGroup extends FoundationElement {
 }