+++ libs/components/src/lib/text-anchor/text-anchor.template.ts
@@ -65,5 +65,6 @@ export const textAnchorTemplate: (
       ${() => focusTemplate}
       ${x => affixIconTemplate(x.icon)}
       ${x => x.text}
+			${x => x.getBodyTemplate?.() ?? ''}
     </a>
 `;};